fix rosu

+++ RosuPP/src/beatmap/attributes.rs
@@ -9,7 +9,7 @@ use mode::Mode;
 use rosu_mods::{GameMods, GameModsIntermode};
 
 /// Summary struct for a [`Beatmap`]'s attributes.
-#[derive(Clone, Debug, PartialEq)]
+#[derive(Clone, Copy, Debug, PartialEq)]
 #[repr(C)]
 #[ffi_type]
 pub struct BeatmapAttributes {
@@ -139,11 +139,11 @@ impl BeatmapAttributesBuilder {
     #[ffi_service_method(on_panic = "undefined_behavior")]
     pub fn get_clock_rate(&mut self) -> f64 {
         if let Some(mods) = self.mods.as_ref() {
-            return f64::from(mods.clock_rate().unwrap_or(1.0))
+            return mods.clock_rate().unwrap_or(1.0)
         }
         
         if let Some(mods_intermode) = self.mods_intermode.as_ref() {
-            return f64::from(mods_intermode.legacy_clock_rate())
+            return mods_intermode.legacy_clock_rate()
         }
 
         1.0
